+++ CHANGES.rst
@@ -1,20 +1,41 @@
 Change Log
 ----------
 
+1.3.0
+~~~~~~~
+ 
+* Fix bulk edit autocomplete
+
+* Added 25 as default value of deprecation_rate
+
+* Created a method in API to retrieve warehouses
+
+* Added fields like venture_id, is_blade, cores_count, power_consumption, height_of_device and warehouse_id to get_assets API
+
+* Added fields like power_consumption and height_of_device to AssetModel model
+
+* Moved category from Asset model to AssetModel model
+
+* Added cores_count method as property to Asset model
+
+
 1.2.13
 ~~~~~~~
 
 * fixes of Discovered column. Also it shows now on csv reports.
+
 
 1.2.12
 ~~~~~~~
 
 * Improved the csv exporting system
 
+
 1.2.11
 ~~~~~~~
 
 * Basing deprecation on invoice date instead of delivery date
+
 
 1.2.10
 ~~~~~~~~~~~
@@ -22,6 +43,7 @@ Change Log
 * Pricing api uses only devices that existed on given date
 
 * Pricing api can use forced deprecation
+
 
 1.2.9
 ~~~~
@@ -31,6 +53,7 @@ Change Log
 * Dynamically requiring 'slots' for blade categories
 
 * Fixed unit tests
+
 
 1.2.7
 ~~~~~